Change Position to float format

+++ src/com/Position.java
@@ -1,14 +1,14 @@
 package com;
 
 public class Position {
-	public int x=0, y=0; 
-	public Position(int x, int y) {
+	public float x=0, y=0; 
+	public Position(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 	@Override
 	public String toString()
 	{
-		return "X:" +  Integer.toString(x) + " Y:" + Integer.toString(y);
+		return "X:" +  Float.toString(x) + " Y:" + Float.toString(y);
 	}
 }
